pyproject.toml
[project]
name = "arcreel"
version = "0.7.2"
description = "AI video generation workspace"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.51",
    "ffmpeg-python>=0.2.0",
    "fastapi>=0.135.1",
    "google-genai>=1.65.0",
    "google-adk>=1.32,<1.33",
    "Pillow>=12.1.1",
    "pydantic>=2.12.5",
    "PyYAML>=6.0.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
    "pyjwt>=2.12.1",
    "sqlalchemy[asyncio]>=2.0.48",
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
    "pwdlib[argon2]>=0.3.0",
    "volcengine-python-sdk[ark]>=5.0.17",
    "openai>=2.30.0",
    "openai-agents>=0.1,<0.2",
    "xai-sdk>=1.8.2",
    "pyjianyingdraft>=0.2.6",
    "instructor>=1.14.5",
]

[dependency-groups]
dev = ["pytest>=9.0.2", "pytest-asyncio>=1.3.0", "pytest-cov>=7.0.0", "ruff>=0.15.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib", "server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["lib", "server"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    'if __name__ == "__main__":',
    "if __name__ == '__main__':",
]

[tool.ruff]
line-length = 120
exclude = [".worktrees", ".claude/worktrees"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E402", "E501"]
